refactor(s2s): purge legacy SID / TS6 debt — node_id is the single mesh identity
Clean-room S2S has no TS6 server-id concept. The sovereign node_id (u64) is now
the ONE mesh identity: it keys server_registry, is the CRDT replica lane
(ReplicaId widened u16->u64), and the handshake/gossip identity.

- server_registry: keyed by NodeId (was Sid=u16); NodeInfo/Node use node_id;
  validNode replaces validSid; MAP shows node_id.
- channel_crdt: ReplicaId u16 -> u64 (replica lane = node id).
- s2s_peer: dropped Sid/local_sid/remote_sid + the handshake sid field;
  registry keyed by node_id; remoteSid()->remoteNodeId(); unknown-remote
  adoption preserved.
- s2s_link: dropped local_sid; CRDT replica = local_node_id; remoteNodeId().
- server.zig: dropped config.server_id (node_id only).
- Removed orphan legacy proto/suimyaku_peer.zig (TS6-style 3-char SIDs,
  superseded by s2s_peer); fixed peer_link doc refs.
- serverinfo: dropped the TS6 'TS<n>ow <sid>' tail + sid/ts_version from
  RPL_VERSION (351) — clean RFC-shaped version line.

All tests green, fmt clean, cross-compiles.

+++ src/substrate/suimyaku/channel_crdt.zig
@@ -12,7 +12,7 @@ pub const Hlc = clock.Hlc;
 pub const VersionVector = clock.VersionVector;
 pub const WireFamily = delta_codec.EntityFamily.channel_modes;
 
-pub const ReplicaId = u16;
+pub const ReplicaId = u64;
 pub const MemberId = u64;
 
 pub const Dot = struct {
